+++ services/townService/src/requestHandlers/PostCoveyTownRequestHandlers.ts
@@ -57,43 +57,61 @@ export interface CommentUpdateRequest {
     comment : Comment
 }
 
-const postTownController = new PostCoveyTownController("testTown", true);
-postTownController.addPlayer(new Player('test'));
+export async function postCreateHandler(_requestData : PostCreateRequest): Promise<ResponseEnvelope<Record<string, null>>> {
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
 
-export async function postCreateHandler(_requestData : PostCreateRequest): Promise<ResponseEnvelope<Post>> {
-    // const townsStore = PostCoveyTownStore.getInstance();
-    // const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
-
-    // if (!townController) {
-    //     return {
-    //         isOK: false,
-    //         response: undefined,
-    //         message: 'Unable to create post' 
-    //     }
-    // }
+    if (!postTownController) {
+        return {
+            isOK: false, 
+            response: {}, 
+            message: "Unable to create post",
+        };
+    }
 
     const post = _requestData.post;
     const result = await postTownController.createPost(post);
     return {
         isOK: true,
-        response: result,
+        response: {},
         message: !result ? 'Invalid password. Please double check your town update password.' : undefined,
     };
 }
 
-export async function postGetHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<Post>> {
-    
-    const postID = _requestData.postID;
-    const result = await postTownController.getPost(postID);
+export async function postGetHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<Post | Object>> {
 
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const postID = _requestData.postID;
+
+    if (!postTownController) {
+        return {
+            isOK: false,
+            response: {},
+            message: 'Unable to get post.'
+        };
+    }
+
+    const result = await postTownController.getPost(postID);
     return {
         isOK: true,
         response: result,
         message: !result ? 'Invalid password. Please double check your town update password.' : undefined,
     };
+
 }
 
-export async function postGetAllIDInTownHandler(_requestData : PostGetIdInTownRequest) : Promise<ResponseEnvelope<string[]>> {
+export async function postGetAllIDInTownHandler(_requestData : PostGetIdInTownRequest) : Promise<ResponseEnvelope<string[] | Object>> {
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+
+    if (!postTownController) {
+        return {
+            isOK: false,
+            response: {},
+            message: 'Unable to get all Post IDs in Town.'
+        };
+    }
     const result : string[] = await postTownController.getAllPostInTown();
 
     return {
@@ -103,11 +121,15 @@ export async function postGetAllIDInTownHandler(_requestData : PostGetIdInTownRe
     };
 }
 
-export async function postDeleteHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<Post>> {
-    if (!postTownController?.getSessionByToken(_requestData.sessionToken)){
+export async function postDeleteHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<Post | Object>> {
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
         return {
           isOK: false, 
-          response: undefined, 
+          response: {}, 
           message: `Unable to delete post with post ID ${_requestData.postID} in town ${_requestData.coveyTownID}`,
         };
     }
@@ -122,9 +144,20 @@ export async function postDeleteHandler(_requestData : PostGetRequest) : Promise
     };
 }
 
-export async function postUpdateHandler(_requestData : PostUpdateRequest) : Promise<ResponseEnvelope<Post>> {
+export async function postUpdateHandler(_requestData : PostUpdateRequest) : Promise<ResponseEnvelope<Post | Object>> {
     const postID = _requestData.postID;
     const post = _requestData.post;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to update post with post ID ${_requestData.postID} in town ${_requestData.coveyTownID}`,
+        };
+    }
 
     const result = await postTownController.updatePost(postID, post, _requestData.sessionToken);
 
@@ -135,9 +168,20 @@ export async function postUpdateHandler(_requestData : PostUpdateRequest) : Prom
     };
 }
 
-export async function commentCreateHandler(_requestData : CommentCreateRequest): Promise<ResponseEnvelope<Comment>> {
+export async function commentCreateHandler(_requestData : CommentCreateRequest): Promise<ResponseEnvelope<Comment | Object>> {
     const comment = _requestData.comment;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+
+    if (!postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to create comment`,
+        };
+    }
     const result = await postTownController.createComment(comment);
+    
     return {
         isOK: true,
         response: result,
@@ -145,9 +189,19 @@ export async function commentCreateHandler(_requestData : CommentCreateRequest):
     };
 }
 
-export async function commentGetHandler(_requestData : CommentGetRequest) : Promise<ResponseEnvelope<Comment>> {
+export async function commentGetHandler(_requestData : CommentGetRequest) : Promise<ResponseEnvelope<Comment | Object>> {
     
     const commentID = _requestData.commentID;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+
+    if (!postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to get comment`,
+        };
+    }
     const result = await postTownController.getComment(commentID);
 
     return {
@@ -157,8 +211,19 @@ export async function commentGetHandler(_requestData : CommentGetRequest) : Prom
     };
 }
 
-export async function commentDeleteHandler(_requestData : CommentGetRequest) : Promise<ResponseEnvelope<Comment>> {
+export async function commentDeleteHandler(_requestData : CommentGetRequest) : Promise<ResponseEnvelope<Comment | Object>> {
     const commentID = _requestData.commentID;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to delete comment with comment ID ${_requestData.commentID} in town ${_requestData.coveyTownID}`,
+        };
+    }
     const result = await postTownController.deleteComment(commentID, _requestData.sessionToken);
 
     return {
@@ -168,9 +233,20 @@ export async function commentDeleteHandler(_requestData : CommentGetRequest) : P
     };
 }
 
-export async function commentUpdateHandler(_requestData : CommentUpdateRequest) : Promise<ResponseEnvelope<Comment>> {
+export async function commentUpdateHandler(_requestData : CommentUpdateRequest) : Promise<ResponseEnvelope<Comment | Object>> {
     const commentID = _requestData.commentID;
     const comment = _requestData.comment;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to update comment with comment ID ${_requestData.commentID} in town ${_requestData.coveyTownID}`,
+        };
+    }
 
     const result = await postTownController.updateComment(commentID, comment, _requestData.sessionToken);
 
@@ -183,6 +259,16 @@ export async function commentUpdateHandler(_requestData : CommentUpdateRequest) 
 
 export async function fileGetHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<any>> {
     const postID = _requestData.postID;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+
+    if (!postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to get file`,
+        };
+    }
     const result = await postTownController.getFile(postID);
 
     return {
@@ -194,6 +280,18 @@ export async function fileGetHandler(_requestData : PostGetRequest) : Promise<Re
 
 export async function fileDeleteHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<any>> {
     const postID = _requestData.postID;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to delete file with post ID ${_requestData.postID} in town ${_requestData.coveyTownID}`,
+        };
+    }
+
     const result = await postTownController.deleteFile(postID, _requestData.sessionToken);
 
     return {
